refactor: gateway_port infrastructure/ports/ → core/gateway/port.py
ports-migrate: 4 소비자 갱신. 행동 변경 없음.
