+++ core/cli/__init__.py
@@ -1575,7 +1575,7 @@ def serve(
 
     # Wire AgenticLoop as gateway processor
     from core.agent.conversation import ConversationContext
-    from core.infrastructure.ports.gateway_port import get_gateway
+    from core.gateway.port import get_gateway
 
     gateway = get_gateway()
     if gateway is None:
